requirements.txt
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
awscli==1.25.36
beautifulsoup4==4.11.1
boto3==1.24.36
botocore==1.27.36
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.2
cssselect==1.1.0
cycler==0.11.0
docutils==0.16
et-xmlfile==1.1.0
filelock==3.6.0
Fiona==1.8.21
fonttools==4.33.3
geopandas==0.10.2
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.5.0
itemloaders==1.0.4
jmespath==1.0.0
kiwisolver==1.4.2
lxml==4.8.0
matplotlib==3.5.2
munch==2.5.0
numpy==1.22.3
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.2
parsel==1.6.0
Pillow==9.1.0
Protego==0.2.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyproj==3.3.1
pyshp==2.3.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
rsa==4.7.2
s3transfer==0.6.0
scipy==1.8.0
Scrapy==2.6.1
seaborn==0.11.2
selenium==4.1.3
service-identity==21.1.0
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tldextract==3.2.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.4.0
typing_extensions==4.2.0
urllib3==1.26.9
w3lib==1.22.0
wsproto==1.1.0
XlsxWriter==3.0.3
zope.interface==5.4.0
